도메인 클래스에 대한 중복 제거 및 리팩토링 - 계속 사용되는 컬럼,함수 중복 제거
