+++ src/main/java/com/codessquad/qna/repository/Question.java
@@ -6,31 +6,23 @@ import com.codessquad.qna.util.PathUtil;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.commons.lang3.ObjectUtils;
-
 import javax.persistence.*;
-import java.time.LocalDate;
 
 @Entity
 @Getter
-public class Question {
-    @Id
-    @GeneratedValue
-    private Long id;
+@Setter
+public class Question extends AbstractEntity{
 
-    @Setter
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
     private User writer;
 
-    @Setter
     @Column(nullable = false)
     private String title;
 
-    @Setter
     @Lob
     @Column(nullable = false)
     private String contents;
-    private LocalDate createdAt = LocalDate.now();
     private Boolean deleted = false;
 
     public Question () {}
