+++ src/main/java/com/codessquad/qna/repository/User.java
@@ -10,32 +10,24 @@ import org.apache.commons.lang3.ObjectUtils;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 
 
 @Entity
 @Getter
-public class User {
-    @Id
-    @GeneratedValue
-    private Long id;
+@Setter
+public class User extends AbstractEntity{
 
     @Column(nullable = false, length = 20, unique = true)
-    @Setter
     private String userId;
 
     @Column(nullable = false)
-    @Setter
     @JsonIgnore
     private String password;
 
     @Column(nullable = false)
-    @Setter
     private String name;
 
     @Column(nullable = false)
-    @Setter
     private String email;
 
     public void update(User updateData, String currentPassword) {
@@ -59,30 +51,5 @@ public class User {
         boolean emailIsExist = ObjectUtils.isNotEmpty(user.email);
 
         return userIdIsExist && nameIsExist && passwordIsExist && emailIsExist;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0: id.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        User other = (User) obj;
-        if (id == null) {
-            if (other.id != null)
-                return false;
-        } else if (!id.equals(other.id))
-            return false;
-        return true;
     }
 }
